Move admin classes, remove function prefixes
Fix autoloader logic to account for files in sub-directories
