+++ views/screen-settings.php
@@ -39,7 +39,7 @@ namespace SatisPress;
 	"repositories": [
 		{
 			"type": "composer",
-			"url": "<?php echo esc_url( satispress_get_packages_permalink( [ 'base' => true ] ) ); ?>"
+			"url": "<?php echo esc_url( get_packages_permalink( [ 'base' => true ] ) ); ?>"
 		}
 	]
 }</code></pre>
